backend/requirements.txt
Flask
Flask-CORS
Flask-RESTful
pandas
numpy
pymc
arviz
scipy
python-dotenv
gunicor
joblib